Update python checks workflow to flake8_py3
Updated the continuous integration job from 'flake8-lint' to 'flake8_py3' in the Python checks workflows. Also updated the version of Python used by the workflow to 3.11 and changed the 'flake8' action repo used. Added a requirement to match the check name with the job name, and integrated GitHub Token as environment variable.